Move components into lib folder
