+++ lib/components/Chart.jsx
@@ -35,125 +35,5 @@
  */
 
 import React from 'react';
-import PropTypes from 'prop-types';
-import { defaults, Line } from 'react-chartjs-2';
 
-defaults.global.tooltips.enabled = false;
-defaults.global.legend.display = false;
-defaults.global.animation.duration = 500;
-
-const labels = {
-    continuous: {
-        index: Array.from(Array(83).keys()),
-        ticks: v => ((v % 2) === 1 ? 2400 + ((v - 1)) : ''),
-    },
-    separated: {
-        index: Array.from(Array(163).keys()),
-        ticks: v => ((v % 4) === 1 ? 2400 + ((v - 1) / 2) : ''),
-    },
-};
-const gridColors = [];
-for (let i = 0; i < 82; i += 1) {
-    gridColors.push('rgba(0, 0, 0, 0.05)');
-    gridColors.push('rgba(0, 0, 0, 0)');
-}
-const kMinRSSIValue = -127;
-const arrayConverter = (arr, separate) => {
-    const a = [kMinRSSIValue];
-    arr.forEach(v => {
-        a.push(v);
-        if (separate) {
-            a.push(kMinRSSIValue);
-        }
-    });
-    a.push(kMinRSSIValue);
-    return a;
-};
-
-const Chart = props => {
-    const {
-        rssi, rssiMax, animationDuration, yMin, yMax, separateFrequencies,
-    } = props;
-    const sep = separateFrequencies ? 'separated' : 'continuous';
-    const rssiData = arrayConverter(rssi, separateFrequencies);
-    const rssiMaxData = arrayConverter(rssiMax, separateFrequencies);
-    const chartData = {
-        labels: labels[sep].index,
-        datasets: [{
-            label: 'rssi',
-            borderColor: 'rgba(79, 140, 196, 1)',
-            backgroundColor: 'rgba(79, 140, 196, 0.3)',
-            borderWidth: 1,
-            fill: true,
-            data: rssiData,
-            xAxisID: 'x-freq-1',
-            pointRadius: 0,
-        }, {
-            label: 'rssiMax',
-            borderColor: 'rgba(179, 40, 96, 0.3)',
-            borderWidth: 1,
-            fill: false,
-            data: rssiMaxData,
-            xAxisID: 'x-freq-1',
-            pointRadius: 0,
-        }],
-    };
-
-    const chartOptions = {
-        animation: { duration: animationDuration },
-        scales: {
-            xAxes: [{
-                id: 'x-freq-1',
-                type: 'category',
-                ticks: {
-                    callback: labels[sep].ticks,
-                    minRotation: 90,
-                    autoSkipPadding: 5,
-                },
-                scaleLabel: {
-                    display: true,
-                    labelString: 'MHz',
-                },
-                gridLines: {
-                    color: gridColors,
-                },
-            }],
-            yAxes: [{
-                type: 'linear',
-                min: -yMax,
-                max: -yMin,
-                ticks: {
-                    callback: v => yMax + v + yMin,
-                    min: -yMax,
-                    max: -yMin,
-                },
-                scaleLabel: {
-                    display: true,
-                    labelString: 'RSSI',
-                },
-            }],
-        },
-    };
-
-    return <Line data={chartData} options={chartOptions} />;
-};
-
-Chart.propTypes = {
-    rssi: PropTypes.arrayOf(Number),
-    rssiMax: PropTypes.arrayOf(Number),
-    animationDuration: PropTypes.number,
-    yMin: PropTypes.number,
-    yMax: PropTypes.number,
-    separateFrequencies: PropTypes.bool,
-};
-
-Chart.defaultProps = {
-    rssi: [],
-    rssiMax: [],
-    animationDuration: 500,
-    yMin: -110,
-    yMax: -20,
-    separateFrequencies: false,
-};
-
-export default Chart;
+export default () => <span>FIXME app Chart</span>;
