+++ lib/components/SidePanel.jsx
